feat(react router): move ui components to a more logical structure
