+++ apps/web/app/routes/battleship.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import type { MetaFunction } from 'react-router';
 import init from '@aamathews23/battleship-web';
-import { Hero } from '@/components/Hero';
+import { Hero } from '@/components/global/Hero';
 import { Board } from '@/components/battleship/Board';
 import { Spinner } from '@/components/ui/Spinner';
 import { useBattleshipStore } from '@/stores/battleship';
